requirements.txt
asgiref==3.2.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.0.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
channels==2.4.0
channels-redis==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.7
cos-python-sdk-v5==1.9.23
crcmod==1.7
cryptography==40.0.1
cycler==0.11.0
daphne==2.5.0
Django==3.1
django-cors-headers==3.14.0
filelock==3.10.0
Flask==2.2.3
fonttools==4.39.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
incremental==22.10.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
mysqlclient==2.1.1
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
paramiko==2.7.2
Pillow==9.4.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.1.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
redis==4.5.3
requests==2.28.2
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
sqlparse @ file:///C:/b/abs_83jn6owfpc/croot/sqlparse_1668793499314/work
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
websockets==10.4
Werkzeug==2.2.3
wincertstore==0.2
xmltodict==0.13.0
zipp==3.15.0
zope.interface==6.0
